Move lapack_connector.h and lapack_wrapper.h to module_external
