+++ source/source_base/module_grid/batch.cpp
@@ -5,7 +5,7 @@
 #include <iterator>
 
 #include "source_base/module_external/blas_connector.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 
 namespace {
