+++ source/source_base/module_mixing/broyden_mixing.cpp
@@ -1,6 +1,6 @@
 #include "broyden_mixing.h"
 
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/memory.h"
 #include "source_base/module_container/base/third_party/blas.h"
 #include "source_base/timer.h"
