+++ source/source_base/module_mixing/pulay_mixing.cpp
@@ -1,6 +1,6 @@
 #include "pulay_mixing.h"
 
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/memory.h"
 #include "source_base/timer.h"
 #include "source_base/tool_title.h"
