+++ source/source_esolver/esolver_ks_lcao_tddft.cpp
@@ -12,7 +12,7 @@
 //--------------temporary----------------------------
 #include "source_base/module_external/blas_connector.h"
 #include "source_base/global_function.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/scalapack_connector.h"
 #include "source_estate/module_charge/symmetry_rho.h"
 #include "source_estate/module_dm/cal_dm_psi.h"
