+++ source/source_estate/module_dm/cal_edm_tddft.cpp
@@ -1,6 +1,6 @@
 #include "cal_edm_tddft.h"
 
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/scalapack_connector.h"
 namespace elecstate
 {
