+++ source/source_hsolver/diago_elpa_native.cpp
@@ -3,7 +3,7 @@
 #include "source_base/module_external/blacs_connector.h"
 #include "source_base/global_variable.h"
 #include "source_io/module_parameter/parameter.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/scalapack_connector.h"
 #include "source_base/timer.h"
 #include "source_base/tool_quit.h"
