+++ source/source_hsolver/diago_iter_assist.cpp
@@ -4,7 +4,7 @@
 #include "source_base/complexmatrix.h"
 #include "source_base/constants.h"
 #include "source_base/global_variable.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/module_device/device.h"
 #include "source_base/parallel_reduce.h"
 #include "source_base/timer.h"
