+++ source/source_hsolver/diago_lapack.cpp
@@ -5,7 +5,7 @@
 #include "diago_lapack.h"
 
 #include "source_base/global_variable.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/timer.h"
 #include "source_base/tool_quit.h"
 #include "source_pw/module_pwdft/global.h"
