+++ source/source_hsolver/kernels/test/math_dngvd_test.cpp
@@ -1,5 +1,5 @@
 #include "source_base/complexmatrix.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/module_device/memory_op.h"
 #include "source_hsolver/kernels/dngvd_op.h"
 #include "source_base/kernels/math_kernel_op.h"
