+++ source/source_hsolver/test/diago_bpcg_test.cpp
@@ -1,5 +1,5 @@
 #include "source_base/inverse_matrix.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_pw/module_pwdft/structure_factor.h"
 #include "source_psi/psi.h"
 #include "source_hamilt/hamilt.h"
