+++ source/source_hsolver/test/diago_cg_test.cpp
@@ -3,7 +3,7 @@
 #include "source_io/module_parameter/parameter.h"
 #undef private
 #include "source_base/inverse_matrix.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_pw/module_pwdft/structure_factor.h"
 #include "source_psi/psi.h"
 #include "source_hamilt/hamilt.h"
