+++ source/source_hsolver/test/diago_david_real_test.cpp
@@ -5,7 +5,7 @@
 #include "source_psi/psi.h"
 #include"gtest/gtest.h"
 #include "source_base/inverse_matrix.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_basis/module_pw/test/test_tool.h"
 #include"mpi.h"
