+++ source/source_hsolver/test/diago_elpa_utils.h
@@ -1,5 +1,5 @@
 #include "source_base/module_external/blacs_connector.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/scalapack_connector.h"
 #include "mpi.h"
