+++ source/source_hsolver/test/diago_mock.h
@@ -1,5 +1,5 @@
 #include<random>
-#include "../../source_base/lapack_connector.h"
+#include "../../source_base/module_external/lapack_connector.h"
 #include "../../source_base/module_external/blas_connector.h"
 #include "mpi.h"
 #include "source_base/parallel_reduce.h"
