+++ source/source_lcao/module_lr/utils/lr_util.cpp
@@ -1,6 +1,6 @@
 #include "source_base/constants.h"
 #include "lr_util.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/scalapack_connector.h"
 namespace LR_Util
 {
