+++ source/source_lcao/module_ri/Inverse_Matrix.hpp
@@ -7,7 +7,7 @@
 #define INVERSE_MATRIX_HPP
 
 #include "Inverse_Matrix.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 
 #include <cassert>
