+++ source/source_lcao/module_rt/evolve_psi.cpp
@@ -2,7 +2,7 @@
 
 #include "band_energy.h"
 #include "middle_hamilt.h"
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/module_container/ATen/kernels/blas.h"   // cuBLAS handle
 #include "source_base/module_container/ATen/kernels/lapack.h" // cuSOLVER handle
 #include "source_base/scalapack_connector.h"
