+++ source/source_lcao/module_rt/middle_hamilt.cpp
@@ -1,6 +1,6 @@
 #include "middle_hamilt.h"
 
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/module_container/ATen/kernels/blas.h"
 #include "source_base/module_device/memory_op.h" // memory operations
 #include "source_base/scalapack_connector.h"
