+++ source/source_lcao/module_rt/solve_propagation.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/scalapack_connector.h"
 #include "source_pw/module_pwdft/global.h"
