+++ source/source_lcao/module_rt/upsi.cpp
@@ -1,6 +1,6 @@
 #include "upsi.h"
 
-#include "source_base/lapack_connector.h"
+#include "source_base/module_external/lapack_connector.h"
 #include "source_base/module_container/ATen/kernels/blas.h"
 #include "source_base/scalapack_connector.h"
