+++ source/source_relax/bfgs.h
@@ -5,7 +5,7 @@
 #include <tuple> 
 #include<algorithm>
 #include<cmath>
-#include"source_base/lapack_connector.h"
+#include"source_base/module_external/lapack_connector.h"
 #include "source_base/matrix.h"
 #include "source_base/matrix3.h"
 #include "source_cell/unitcell.h"
